Fix code review suggestions
