+++ native/src/main/java/io/ballerina/stdlib/grpc/builder/BallerinaFileBuilder.java
@@ -32,6 +32,7 @@ import io.ballerina.stdlib.grpc.builder.stub.StubFile;
 import io.ballerina.stdlib.grpc.builder.syntaxtree.SyntaxTreeGenerator;
 import io.ballerina.stdlib.grpc.builder.syntaxtree.components.Class;
 import io.ballerina.stdlib.grpc.exception.CodeBuilderException;
+import io.ballerina.stdlib.grpc.protobuf.descriptor.DescriptorMeta;
 import org.apache.commons.lang3.StringUtils;
 import org.ballerinalang.formatter.core.Formatter;
 import org.ballerinalang.formatter.core.FormatterException;
@@ -45,8 +46,8 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -57,8 +58,8 @@ import java.util.stream.Collectors;
  * Class is responsible of generating the ballerina stub which is mapping proto definition.
  */
 public class BallerinaFileBuilder {
-    private byte[] rootDescriptor;
-    private final Set<byte[]> dependentDescriptors;
+    private final DescriptorMeta rootDescriptor;
+    private final Set<DescriptorMeta> dependentDescriptors;
     private String balOutPath;
 
     // Proto file extension
@@ -66,32 +67,37 @@ public class BallerinaFileBuilder {
     public static Map<String, String> enumDefaultValueMap = new HashMap<>();
     public static Map<String, Boolean> dependentValueTypeMap = new HashMap<>();
     public static Map<String, Class> streamClassMap;
-    private Map<String, List<String>> unusedImports;
 
-    public BallerinaFileBuilder(byte[] rootDescriptor, Set<byte[]> dependentDescriptors) {
-        setRootDescriptor(rootDescriptor);
+    public BallerinaFileBuilder(DescriptorMeta rootDescriptor, Set<DescriptorMeta> dependentDescriptors) {
+        this.rootDescriptor = rootDescriptor;
         this.dependentDescriptors = dependentDescriptors;
         streamClassMap = new HashMap<>();
     }
 
-    public BallerinaFileBuilder(byte[] rootDescriptor, Set<byte[]> dependentDescriptors, String balOutPath) {
-        setRootDescriptor(rootDescriptor);
+    public BallerinaFileBuilder(DescriptorMeta rootDescriptor, Set<DescriptorMeta> dependentDescriptors,
+                                String balOutPath) {
+        this.rootDescriptor = rootDescriptor;
         this.dependentDescriptors = dependentDescriptors;
         this.balOutPath = balOutPath;
         streamClassMap = new HashMap<>();
     }
 
-    public void build(String mode, Map<String, List<String>> unusedImports) throws CodeBuilderException {
-        this.unusedImports = unusedImports;
+    public void build(String mode) throws CodeBuilderException {
         // compute dependent descriptor source code.
-        for (byte[] descriptorData : dependentDescriptors) {
+        for (byte[] descriptorData : getDependentDescriptorSet(dependentDescriptors)) {
             computeSourceContent(descriptorData, null, false);
         }
         // compute root descriptor source code.
-        computeSourceContent(rootDescriptor, mode, true);
+        computeSourceContent(rootDescriptor.getDescriptor(), mode, true);
     }
 
     private void computeSourceContent(byte[] descriptor, String mode, boolean isRoot) throws CodeBuilderException {
+        Map<String, List<String>> unusedImports = new HashMap<>();
+        if (rootDescriptor.getUnusedImports().size() > 0) {
+            unusedImports.put(rootDescriptor.getProtoName(), rootDescriptor.getUnusedImports());
+        }
+        unusedImports.putAll(getUnusedImportsInDependentDescriptorSet(dependentDescriptors));
+
         try (InputStream targetStream = new ByteArrayInputStream(descriptor)) {
             // define extension register and register custom option
             ExtensionRegistry extensionRegistry = ExtensionRegistry.newInstance();
@@ -137,14 +143,14 @@ public class BallerinaFileBuilder {
                 }
             }
 
-            if (descriptor == rootDescriptor || serviceDescriptorList.size() > 0) {
+            if (descriptor == rootDescriptor.getDescriptor() || serviceDescriptorList.size() > 0) {
                 // Add root descriptor.
                 Descriptor rootDesc = Descriptor.newBuilder(descriptor).build();
                 stubRootDescriptor = rootDesc.getData();
                 descriptors.add(rootDesc);
 
                 // Add dependent descriptors.
-                for (byte[] descriptorData : dependentDescriptors) {
+                for (byte[] descriptorData : getDependentDescriptorSet(dependentDescriptors)) {
                     Descriptor dependentDescriptor = Descriptor.newBuilder(descriptorData).build();
                     descriptors.add(dependentDescriptor);
                 }
@@ -272,7 +278,21 @@ public class BallerinaFileBuilder {
         }
     }
 
-    private void setRootDescriptor(byte[] rootDescriptor) {
-        this.rootDescriptor = Arrays.copyOf(rootDescriptor, rootDescriptor.length);
+    private Set<byte[]> getDependentDescriptorSet(Set<DescriptorMeta> descriptorMetaSet) {
+        Set<byte[]> dependentDescriptorSet = new HashSet<>();
+        for (DescriptorMeta descriptorMeta : descriptorMetaSet) {
+            dependentDescriptorSet.add(descriptorMeta.getDescriptor());
+        }
+        return dependentDescriptorSet;
+    }
+
+    private Map<String, List<String>> getUnusedImportsInDependentDescriptorSet(Set<DescriptorMeta> descriptorMetaSet) {
+        Map<String, List<String>> unusedImports = new HashMap<>();
+        for (DescriptorMeta descriptorMeta : descriptorMetaSet) {
+            if (descriptorMeta.getUnusedImports().size() > 0) {
+                unusedImports.put(descriptorMeta.getProtoName(), descriptorMeta.getUnusedImports());
+            }
+        }
+        return unusedImports;
     }
 }
